requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
env==0.1.0
filelock==3.13.1
idna==3.6
mysqlclient==2.2.4
pillow==10.3.0
platformdirs==4.2.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.1
virtualenv==20.25.1
